+++ CHANGELOG.md
@@ -4,6 +4,8 @@
 
 <!-- ✨ You do not need to add a pull request reference or an author, this will be added automatically by CI. ✨ -->
 
+ - Add support for rendering [Mermaid diagrams](https://mermaid.js.org/) by passing `--mermaid`.
+   ([#525](https://github.com/mitmproxy/pdoc/pull/525), @thearchitector, @mhils)
  - Add additional Jinja2 blocks to allow a more fine-grained customization of the menu.
    ([#521](https://github.com/mitmproxy/pdoc/pull/521), @mikkelakromann)
  - Fix a crash in pdoc 13.0.0 when `__init__.py` is passed as a file to pdoc.